requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Deprecated==1.2.13
Django==3.2.13
django-extensions==3.1.5
django-redis==5.2.0
djangorestframework==3.13.1
greenlet==1.1.2
jmespath==1.0.1
mysqlclient==2.1.0
numpy==1.23.0
packaging==21.3
pandas==1.4.2
psycopg2-binary==2.9.3
py-dbutils==0.4.28
pyinstrument==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.3
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
tzdata==2022.1
wrapt==1.14.1
